Linux Unhatched
This was a great introduction to the Linux command line, building a solid foundation in navigating the terminal, managing files and directories, and understanding essential commands. It’s a key step in strengthening my core skills for cybersecurity, where Linux knowledge is crucial for tasks like system monitoring, scripting, and incident response.